requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cleanup==5.1.0
django-crispy-forms==1.11.0
Pillow==8.1.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
validate-email==1.3
